requirements.txt
bcrypt==3.1.4
blinker==1.4
beautifulsoup4==4.6.0
cfscrape==1.9.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.3.2
influxdb==4.1.1
python-magic==0.4.15
passwordmeter==0.1.8
Pillow==4.3.0
PyMySQL==0.7.11
python-tumblpy==1.1.4
raven==6.3.0
requests==2.18.4
simple-crypt==4.1.7
tweepy==3.5.0
flask_migrate==2.1.1
